We started adding a new agent, which required a few changes:
The package added agent.ours includes the classes that were added. 
Policy.java : An interface with a method getAction that returns an action for a given state.
EarlyGameOverException.java: An exception that will arise if in the case that the game ends before N turns are over
PolicyAgent.java: An agent that uses a given policy to play the game
Test.java: the main file that follows the policy evaluation and returns the action-value function (the average of rewards after M games)

However there is a problem with these additions. The program seems to get caught in the getRandomState method for a reason that could not be deduced.  (David Sz thinks it's because the randompolicy we use to generate the random states dies way too early. Fix it)