THIS COMMIT CONTAINS THINGS THAT DON'T RUN / WORK YET. All the partial work done is committed here before returning the laptop.
The bugs in the evaluatePolicy method have been fixed in Test.java although the returned rewards are very low (most of them are 0.0 with the exception of a few cases) as excepted.
WeightPolicy.java --> Follows the PEGASOS algorithm. Collects and updates the weights to a given policy--> Not yet finished. Needs further modifications
Everything else is pretty much the same.